Move ViewManager to its own module
