+++ bokehjs/src/lib/embed/standalone.ts
@@ -2,7 +2,7 @@ import type {Document} from "../document"
 import {RootAddedEvent, RootRemovedEvent, TitleChangedEvent} from "../document"
 import type {HasProps} from "../core/has_props"
 import type {View} from "../core/view"
-import {ViewManager} from "../core/view"
+import {ViewManager} from "../core/view_manager"
 import {DOMView} from "../core/dom_view"
 import {build_view} from "../core/build_views"
 import {isString} from "../core/util/types"
