+++ bokehjs/src/lib/models/callbacks/customjs.ts
@@ -8,7 +8,7 @@ import type {Model} from "../../model"
 import {logger} from "core/logging"
 import type {Dict} from "core/types"
 import {isFunction} from "core/util/types"
-import type {ViewManager} from "core/view"
+import type {ViewManager} from "core/view_manager"
 import {index} from "embed/standalone"
 
 type KV = {[key: string]: unknown}
